Move Widget channel manager to base widget folder
