+++ base-widget/src/WidgetChannelManager.js
